refactor(#741): refactor partnerController
