+++ api/controllers/v1/partner/find-all.js
@@ -1,0 +1,20 @@
+const ControllerService = require('../../../services/ControllerService');
+
+module.exports = (req, res) => {
+  const parameters = {};
+  if (req.param('name')) {
+    parameters.name = {
+      like: `%${req.param('name')}%`,
+    };
+    sails.log.debug(`parameters ${parameters.name.like}`);
+  }
+
+  TGrotto.find(parameters)
+    .sort('id ASC')
+    .exec((err, found) => {
+      const params = {};
+      params.controllerMethod = 'PartnerController.findAll';
+      params.notFoundMessage = 'No partners found.';
+      return ControllerService.treat(req, err, found, params, res);
+    });
+};
